feat: Add VirtualBox serial logging and fix VirtIO 64-bit BAR handling for MMIO regions
- Added serial port configuration (COM1 0x3F8) to VirtualBox VM with output to serial.log file
- Added automatic VM poweroff before configuration changes and tail -f serial.log after VM start
- Fixed VirtIO BAR address/size handling to use uint64_t instead of uint32_t for 64-bit BAR support
- Added 64-bit BAR detection (bits 2:1 = 10) in get_bar_info() to read upper 32 bits from BAR+4
- Fixed virtio_mmio_probe

+++ src/drivers/virtio/virtio_mmio.c
@@ -85,7 +85,7 @@ static void virtio_mmio_add_status(VirtioMmioDevice *dev, uint32_t status_bit)
 /*           API publique                       */
 /* ============================================ */
 
-VirtioMmioDevice *virtio_mmio_probe(uint32_t phys_addr, uint32_t size, uint32_t irq)
+VirtioMmioDevice *virtio_mmio_probe(uint64_t phys_addr, uint64_t size, uint32_t irq)
 {
     KLOG_INFO("VIRTIO_MMIO", "Probing device at phys addr:");
     KLOG_INFO_HEX("VIRTIO_MMIO", "  Address: ", phys_addr);
